Correct output path in exp..py and describe feature importance using xgb scripts
----------------------
Extreme gradient boosting is a method that allows to find feature importance from sparse matrices. In contrast to other methods,
it can handle missigness in data by deciding to which side of the split the missingness belongs. This approach allows the algorithm
to use missing data as part of the dataset to classify the rows.